Add resource consumption and collection

+++ Hive2D/Hive2D/Game/Components/ResourceComponent.swift
@@ -9,9 +9,9 @@
 import GameplayKit
 
 class ResourceComponent: GKComponent {
-    var resources: Int
+    var resources: Float
 
-    init(resources: Int) {
+    init(resources: Float) {
         self.resources = resources
         super.init()
     }
